move github_apps out of enterprise
