+++ enterprise/cmd/frontend/internal/auth/githubappauth/BUILD.bazel
@@ -17,8 +17,6 @@ go_library(
         "//cmd/frontend/graphqlbackend",
         "//enterprise/cmd/frontend/internal/repos/webhooks/resolvers",
         "//enterprise/internal/database",
-        "//enterprise/internal/github_apps/auth",
-        "//enterprise/internal/github_apps/types",
         "//internal/auth",
         "//internal/codeintel",
         "//internal/conf/conftypes",
@@ -27,6 +25,8 @@ go_library(
         "//internal/encryption/keyring",
         "//internal/extsvc",
         "//internal/extsvc/github",
+        "//internal/github_apps/auth",
+        "//internal/github_apps/types",
         "//internal/gqlutil",
         "//internal/observation",
         "//internal/rcache",
@@ -57,11 +57,11 @@ go_test(
     deps = [
         "//cmd/frontend/graphqlbackend",
         "//enterprise/internal/database",
-        "//enterprise/internal/github_apps/store",
-        "//enterprise/internal/github_apps/types",
         "//internal/actor",
         "//internal/database",
         "//internal/encryption",
+        "//internal/github_apps/store",
+        "//internal/github_apps/types",
         "//internal/gitserver",
         "//internal/rcache",
         "//internal/types",
