+++ enterprise/cmd/frontend/internal/batches/resolvers/BUILD.bazel
@@ -56,8 +56,6 @@ go_library(
         "//enterprise/internal/batches/types",
         "//enterprise/internal/batches/types/scheduler/config",
         "//enterprise/internal/database",
-        "//enterprise/internal/github_apps/store",
-        "//enterprise/internal/github_apps/types",
         "//internal/actor",
         "//internal/api",
         "//internal/auth",
@@ -72,6 +70,8 @@ go_library(
         "//internal/extsvc/bitbucketserver",
         "//internal/extsvc/github",
         "//internal/featureflag",
+        "//internal/github_apps/store",
+        "//internal/github_apps/types",
         "//internal/gitserver",
         "//internal/gitserver/gitdomain",
         "//internal/gqlutil",
