+++ enterprise/cmd/gitserver/shared/BUILD.bazel
@@ -12,13 +12,13 @@ go_library(
         "//cmd/gitserver/shared",
         "//enterprise/internal/authz/subrepoperms",
         "//enterprise/internal/database",
-        "//enterprise/internal/github_apps/auth",
         "//internal/authz",
         "//internal/database",
         "//internal/debugserver",
         "//internal/encryption/keyring",
         "//internal/env",
         "//internal/extsvc/github/auth",
+        "//internal/github_apps/auth",
         "//internal/observation",
         "//internal/service",
         "@com_github_sourcegraph_log//:log",
