+++ enterprise/cmd/repo-updater/shared/BUILD.bazel
@@ -19,13 +19,13 @@ go_library(
         "//enterprise/internal/authz",
         "//enterprise/internal/batches",
         "//enterprise/internal/database",
-        "//enterprise/internal/github_apps/auth",
         "//internal/actor",
         "//internal/authz",
         "//internal/conf",
         "//internal/database",
         "//internal/encryption/keyring",
         "//internal/extsvc/github/auth",
+        "//internal/github_apps/auth",
         "//internal/goroutine",
         "//internal/httpcli",
         "//internal/licensing",
