+++ enterprise/cmd/worker/internal/githubapps/worker/BUILD.bazel
@@ -8,9 +8,9 @@ go_library(
     visibility = ["//enterprise/cmd/worker:__subpackages__"],
     deps = [
         "//enterprise/internal/database",
-        "//enterprise/internal/github_apps/auth",
-        "//enterprise/internal/github_apps/types",
         "//internal/extsvc/github",
+        "//internal/github_apps/auth",
+        "//internal/github_apps/types",
         "//internal/goroutine",
         "//lib/errors",
         "@com_github_sourcegraph_log//:log",
@@ -23,8 +23,8 @@ go_test(
     embed = [":worker"],
     deps = [
         "//enterprise/internal/database",
-        "//enterprise/internal/github_apps/store",
-        "//enterprise/internal/github_apps/types",
+        "//internal/github_apps/store",
+        "//internal/github_apps/types",
         "//internal/types",
         "//lib/errors",
         "@com_github_google_go_github_v41//github",
