+++ enterprise/internal/authz/BUILD.bazel
@@ -51,6 +51,7 @@ go_test(
         "@com_github_google_go_cmp//cmp/cmpopts",
         "@com_github_inconshreveable_log15//:log15",
         "@com_github_json_iterator_go//:go",
+        "@com_github_kr_pretty//:pretty",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
     ],
