+++ enterprise/internal/batches/sources/BUILD.bazel
@@ -23,8 +23,6 @@ go_library(
         "//enterprise/internal/batches/sources/gerrit",
         "//enterprise/internal/batches/store",
         "//enterprise/internal/batches/types",
-        "//enterprise/internal/github_apps/auth",
-        "//enterprise/internal/github_apps/store",
         "//internal/api",
         "//internal/conf",
         "//internal/database",
@@ -40,6 +38,8 @@ go_library(
         "//internal/extsvc/github/auth",
         "//internal/extsvc/gitlab",
         "//internal/extsvc/versions",
+        "//internal/github_apps/auth",
+        "//internal/github_apps/store",
         "//internal/gitserver",
         "//internal/gitserver/gitdomain",
         "//internal/gitserver/protocol",
@@ -77,9 +77,6 @@ go_test(
         "//enterprise/internal/batches/sources/gerrit",
         "//enterprise/internal/batches/store",
         "//enterprise/internal/batches/types",
-        "//enterprise/internal/github_apps/auth",
-        "//enterprise/internal/github_apps/store",
-        "//enterprise/internal/github_apps/types",
         "//internal/api",
         "//internal/authz",
         "//internal/conf",
@@ -94,6 +91,9 @@ go_test(
         "//internal/extsvc/github",
         "//internal/extsvc/gitlab",
         "//internal/extsvc/versions",
+        "//internal/github_apps/auth",
+        "//internal/github_apps/store",
+        "//internal/github_apps/types",
         "//internal/gitserver",
         "//internal/gitserver/gitdomain",
         "//internal/gitserver/protocol",
