+++ enterprise/internal/batches/sources/mocks_test.go
@@ -18,7 +18,6 @@ import (
 	diff "github.com/sourcegraph/go-diff/diff"
 	store "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/store"
 	types1 "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/types"
-	store1 "github.com/sourcegraph/sourcegraph/enterprise/internal/github_apps/store"
 	api "github.com/sourcegraph/sourcegraph/internal/api"
 	authz "github.com/sourcegraph/sourcegraph/internal/authz"
 	database "github.com/sourcegraph/sourcegraph/internal/database"
@@ -26,6 +25,7 @@ import (
 	azuredevops "github.com/sourcegraph/sourcegraph/internal/extsvc/azuredevops"
 	bitbucketcloud "github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketcloud"
 	gerrit "github.com/sourcegraph/sourcegraph/internal/extsvc/gerrit"
+	store1 "github.com/sourcegraph/sourcegraph/internal/github_apps/store"
 	gitserver "github.com/sourcegraph/sourcegraph/internal/gitserver"
 	gitdomain "github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	protocol "github.com/sourcegraph/sourcegraph/internal/gitserver/protocol"
