+++ enterprise/internal/batches/store/BUILD.bazel
@@ -34,7 +34,6 @@ go_library(
         "//enterprise/internal/batches/sources/gerrit",
         "//enterprise/internal/batches/store/author",
         "//enterprise/internal/batches/types",
-        "//enterprise/internal/github_apps/store",
         "//internal/actor",
         "//internal/api",
         "//internal/database",
@@ -53,6 +52,7 @@ go_library(
         "//internal/extsvc/github",
         "//internal/extsvc/gitlab",
         "//internal/featureflag",
+        "//internal/github_apps/store",
         "//internal/gitserver/protocol",
         "//internal/metrics",
         "//internal/observation",
