+++ enterprise/internal/batches/syncer/BUILD.bazel
@@ -16,11 +16,11 @@ go_library(
         "//enterprise/internal/batches/state",
         "//enterprise/internal/batches/store",
         "//enterprise/internal/batches/types",
-        "//enterprise/internal/github_apps/store",
         "//internal/api",
         "//internal/batches",
         "//internal/conf",
         "//internal/database",
+        "//internal/github_apps/store",
         "//internal/gitserver",
         "//internal/goroutine",
         "//internal/httpcli",
@@ -46,10 +46,10 @@ go_test(
     deps = [
         "//enterprise/internal/batches/store",
         "//enterprise/internal/batches/types",
-        "//enterprise/internal/github_apps/store",
         "//internal/api",
         "//internal/database",
         "//internal/extsvc",
+        "//internal/github_apps/store",
         "//internal/observation",
         "//internal/timeutil",
         "//internal/types",
