+++ enterprise/internal/repos/BUILD.bazel
@@ -10,13 +10,13 @@ go_test(
     ],
     deps = [
         "//enterprise/internal/database",
-        "//enterprise/internal/github_apps/auth",
-        "//enterprise/internal/github_apps/types",
         "//internal/database",
         "//internal/database/dbtest",
         "//internal/encryption/keyring",
         "//internal/extsvc",
         "//internal/extsvc/github/auth",
+        "//internal/github_apps/auth",
+        "//internal/github_apps/types",
         "//internal/httpcli",
         "//internal/rcache",
         "//internal/repos",
