+++ internal/github_apps/auth/BUILD.bazel
@@ -4,14 +4,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "auth",
     srcs = ["auth.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/github_apps/auth",
-    visibility = ["//enterprise:__subpackages__"],
+    importpath = "github.com/sourcegraph/sourcegraph/internal/github_apps/auth",
+    visibility = ["//:__subpackages__"],
     deps = [
-        "//enterprise/internal/github_apps/store",
         "//internal/encryption",
         "//internal/extsvc/auth",
         "//internal/extsvc/github",
         "//internal/extsvc/github/auth",
+        "//internal/github_apps/store",
         "//internal/httpcli",
         "//internal/rcache",
         "//lib/errors",
@@ -29,10 +29,10 @@ go_test(
         "requires-network",
     ],
     deps = [
-        "//enterprise/internal/github_apps/store",
-        "//enterprise/internal/github_apps/types",
         "//internal/encryption/keyring",
         "//internal/extsvc/auth",
+        "//internal/github_apps/store",
+        "//internal/github_apps/types",
         "//internal/rcache",
         "//schema",
         "@com_github_google_uuid//:uuid",
