+++ internal/github_apps/store/BUILD.bazel
@@ -7,15 +7,15 @@ go_library(
         "mocks_temp.go",
         "store.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/github_apps/store",
-    visibility = ["//enterprise:__subpackages__"],
+    importpath = "github.com/sourcegraph/sourcegraph/internal/github_apps/store",
+    visibility = ["//:__subpackages__"],
     deps = [
-        "//enterprise/internal/github_apps/types",
         "//internal/database/basestore",
         "//internal/database/dbutil",
         "//internal/encryption",
         "//internal/encryption/keyring",
         "//internal/extsvc",
+        "//internal/github_apps/types",
         "//internal/types",
         "//lib/errors",
         "@com_github_keegancsmith_sqlf//:sqlf",
@@ -33,10 +33,9 @@ go_test(
         "requires-network",
     ],
     deps = [
-        "//enterprise/internal/github_apps/types",
-        "//internal/database",
         "//internal/database/basestore",
         "//internal/database/dbtest",
+        "//internal/github_apps/types",
         "//internal/types",
         "@com_github_google_go_github_v41//github",
         "@com_github_keegancsmith_sqlf//:sqlf",
